+++ prefs.js
@@ -1,5 +1,5 @@
 /*
- * Publication Rankings Plugin for Zotero 7
+ * Publication Rankings Plugin for Zotero 7+
  * Default Preferences
  * 
  * Copyright (C) 2025 Ben Stephens
@@ -19,6 +19,9 @@ pref("extensions.publication-rankings.enableABS", true);
 
 // Enable FT50 journal rankings
 pref("extensions.publication-rankings.enableFT50", true);
+
+// Enable VHB journal rankings
+pref("extensions.publication-rankings.enableVHB", true);
 
 // Disable Badges
 pref("extensions.publication-rankings.enableBadges", false);